did a bunch of fixes
fixed some stuff about floating pooint numbers
improved the REPL
made it work but only temporarily. the bug was that define/set returned the value, and if it was a function/exec  it would try to call it which isnt great
output.txt was a debug file that accidentaly got commited. 1.3mb lol